some changes, mainly to the time functions
