+++ src/main.css
@@ -155,24 +155,27 @@ body {
     margin: 0 auto;
 }
 
- #time{
-  margin-left: -36.5px;
-  width: 78px;
-  align-items: center;
-  background-color: rgba(55, 55, 55, 0.5);
-  border-radius: 0 0 5px 5px;
-  border: 2px solid black;
-  border-top: none;
-  padding: 5px;
-  color:#00dd3c;
-  font-family: 'Press Start 2P', cursive;
-  font-size: 10px;
-}
-
 #ingameTime{
     margin-left: -36.5px;
     width: 78px;
     align-items: center;
+    text-align: center;
+    background-color: rgba(55, 55, 55, 0.5);
+    border-radius: 0 0 5px 5px;
+    border: 2px solid black;
+    border-top: none;
+    padding: 5px;
+    color:#00dd3c;
+    font-family: 'Press Start 2P';
+    font-size: 10px;
+}
+
+#timer {
+    visibility: hidden;
+    margin-left: -36.5px;
+    width: 78px;
+    align-items: center;
+    text-align: center;
     background-color: rgba(55, 55, 55, 0.5);
     border-radius: 0 0 5px 5px;
     border: 2px solid black;
